+++ ContentPatcher/Framework/Tokens/ValueProviders/RandomValueProvider.cs
@@ -48,7 +48,7 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders
 
             // get random number for input
             string seedString = input.GetRawArgumentValue("key") ?? input.TokenString.Path;
-            int randomNumber = new Random(unchecked(this.BaseSeed + seedString.GetHashCode())).Next();
+            int randomNumber = new Random(unchecked(this.BaseSeed + this.GetDeterministicHashCode(seedString))).Next();
 
             // choose value
             yield return input.PositionalArgs[randomNumber % input.PositionalArgs.Length];
@@ -87,5 +87,28 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders
 
             return unchecked(daysSinceStart + uniqueId);
         }
+
+        /// <summary>Get a deterministic hash code for a given string.</summary>
+        /// <param name="str">The string to hash.</param>
+        /// <remarks>This ensures that the same hash code is generated across multiple players in multiplayer, so randomization is in sync. Derived from <a href="https://andrewlock.net/why-is-string-gethashcode-different-each-time-i-run-my-program-in-net-core/#a-deterministic-gethashcode-implementation">code by Andrew Lock</a>.</remarks>
+        private int GetDeterministicHashCode(string str)
+        {
+            unchecked
+            {
+                int hash1 = (5381 << 16) + 5381;
+                int hash2 = hash1;
+
+                for (int i = 0; i < str.Length; i += 2)
+                {
+                    hash1 = ((hash1 << 5) + hash1) ^ str[i];
+                    if (i == str.Length - 1)
+                        break;
+                    hash2 = ((hash2 << 5) + hash2) ^ str[i + 1];
+                }
+
+                return hash1 + (hash2 * 1566083941);
+            }
+        }
+
     }
 }